requirements.txt
Pillow>=10.0.0
tkinter-dnd2>=0.3.0
